pyproject.toml
[project]
name = "bootdev-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

[dependency-groups]
dev = [
    "yapf>=0.43.0",
]
